+++ spec/js_routes/options_spec.rb
@@ -13,27 +13,38 @@ describe JsRoutes, "options" do
   let(:_options) { {} }
   let(:_warnings) { true }
 
-  context "when serializer is specified" do
-    # define custom serializer
-    # this is a nonsense serializer, which always returns foo=bar
-    # for all inputs
-    let(:_presetup){ %q(function myCustomSerializer(object, prefix) { return "foo=bar"; }) }
-    let(:_options) { {:serializer => "myCustomSerializer"} }
+  describe "serializer" do
+    context "when specified" do
+      # define custom serializer
+      # this is a nonsense serializer, which always returns foo=bar
+      # for all inputs
+      let(:_presetup){ %q(function myCustomSerializer(object, prefix) { return "foo=bar"; }) }
+      let(:_options) { {:serializer => "myCustomSerializer"} }
 
-    it "should set configurable serializer" do
-      # expect the nonsense serializer above to have appened foo=bar
-      # to the end of the path
-      expect(evaljs(%q(Routes.inboxes_path()))).to eql("/inboxes?foo=bar")
+      it "should set configurable serializer" do
+        # expect the nonsense serializer above to have appened foo=bar
+        # to the end of the path
+        expect(evaljs(%q(Routes.inboxes_path()))).to eql("/inboxes?foo=bar")
+      end
     end
-  end
 
-  context "when serializer is specified, but not function" do
-    let(:_presetup){ %q(var myCustomSerializer = 1) }
-    let(:_options) { {:serializer => "myCustomSerializer"} }
+    context "when specified, but not function" do
+      let(:_presetup){ %q(var myCustomSerializer = 1) }
+      let(:_options) { {:serializer => "myCustomSerializer"} }
 
-    it "should set configurable serializer" do
-      # expect to use default
-      expect(evaljs(%q(Routes.inboxes_path({a: 1})))).to eql("/inboxes?a=1")
+      it "should set configurable serializer" do
+        # expect to use default
+        expect(evaljs(%q(Routes.inboxes_path({a: 1})))).to eql("/inboxes?a=1")
+      end
+    end
+
+    context "when configured in js" do
+      let(:_options) { {:serializer =>%q(function (object, prefix) { return "foo=bar"; })} }
+
+      it "uses JS serializer" do
+        evaljs("Routes.configure({serializer: function (object, prefix) { return 'bar=baz'; }})")
+        expect(evaljs(%q(Routes.inboxes_path({a: 1})))).to eql("/inboxes?bar=baz")
+      end
     end
   end
 
@@ -84,12 +95,13 @@ describe JsRoutes, "options" do
     let(:_options) { {:prefix => "/myprefix/" } }
 
     it "should render routing with prefix" do
-        expect(evaljs("Routes.inbox_path(1)")).to eq("/myprefix#{routes.inbox_path(1)}")
+        expect(evaljs("Routes.inbox_path(1)")).to eq("/myprefix#{test_routes.inbox_path(1)}")
     end
 
     it "should render routing with prefix set in JavaScript" do
-      evaljs("Routes.options.prefix = '/newprefix/'")
-      expect(evaljs("Routes.inbox_path(1)")).to eq("/newprefix#{routes.inbox_path(1)}")
+      evaljs("Routes.configure({prefix: '/newprefix/'})")
+      expect(evaljs("Routes.config().prefix")).to eq("/newprefix/")
+      expect(evaljs("Routes.inbox_path(1)")).to eq("/newprefix#{test_routes.inbox_path(1)}")
     end
 
   end
@@ -99,7 +111,7 @@ describe JsRoutes, "options" do
     let(:_options) { {:prefix => "http://localhost:3000" } }
 
     it "should render routing with prefix" do
-      expect(evaljs("Routes.inbox_path(1)")).to eq(_options[:prefix] + routes.inbox_path(1))
+      expect(evaljs("Routes.inbox_path(1)")).to eq(_options[:prefix] + test_routes.inbox_path(1))
     end
   end
 
@@ -108,12 +120,12 @@ describe JsRoutes, "options" do
     let(:_options) { {:prefix => "/myprefix" } }
 
     it "should render routing with prefix" do
-      expect(evaljs("Routes.inbox_path(1)")).to eq("/myprefix#{routes.inbox_path(1)}")
+      expect(evaljs("Routes.inbox_path(1)")).to eq("/myprefix#{test_routes.inbox_path(1)}")
     end
 
     it "should render routing with prefix set in JavaScript" do
-      evaljs("Routes.options.prefix = '/newprefix'")
-      expect(evaljs("Routes.inbox_path(1)")).to eq("/newprefix#{routes.inbox_path(1)}")
+      evaljs("Routes.configure({prefix: '/newprefix/'})")
+      expect(evaljs("Routes.inbox_path(1)")).to eq("/newprefix#{test_routes.inbox_path(1)}")
     end
 
   end
@@ -123,31 +135,31 @@ describe JsRoutes, "options" do
     let(:_warnings) { nil }
 
     it "should render routing with default_format" do
-      expect(evaljs("Routes.inbox_path(1)")).to eq(routes.inbox_path(1, :format => "json"))
+      expect(evaljs("Routes.inbox_path(1)")).to eq(test_routes.inbox_path(1, :format => "json"))
     end
 
     it "should render routing with default_format and zero object" do
-      expect(evaljs("Routes.inbox_path(0)")).to eq(routes.inbox_path(0, :format => "json"))
+      expect(evaljs("Routes.inbox_path(0)")).to eq(test_routes.inbox_path(0, :format => "json"))
     end
 
     it "should override default_format when spefified implicitly" do
-      expect(evaljs("Routes.inbox_path(1, {format: 'xml'})")).to eq(routes.inbox_path(1, :format => "xml"))
+      expect(evaljs("Routes.inbox_path(1, {format: 'xml'})")).to eq(test_routes.inbox_path(1, :format => "xml"))
     end
 
     it "should override nullify implicitly when specified implicitly" do
-      expect(evaljs("Routes.inbox_path(1, {format: null})")).to eq(routes.inbox_path(1))
+      expect(evaljs("Routes.inbox_path(1, {format: null})")).to eq(test_routes.inbox_path(1))
     end
 
 
     it "shouldn't require the format" do
       pending if Rails.version < "4.0"
-      expect(evaljs("Routes.json_only_path()")).to eq(routes.json_only_path(:format => 'json'))
+      expect(evaljs("Routes.json_only_path()")).to eq(test_routes.json_only_path(:format => 'json'))
     end
   end
 
   it "shouldn't include the format when {:format => false} is specified" do
-    expect(evaljs("Routes.no_format_path()")).to eq(routes.no_format_path())
-    expect(evaljs("Routes.no_format_path({format: 'json'})")).to eq(routes.no_format_path(format: 'json'))
+    expect(evaljs("Routes.no_format_path()")).to eq(test_routes.no_format_path())
+    expect(evaljs("Routes.no_format_path({format: 'json'})")).to eq(test_routes.no_format_path(format: 'json'))
   end
 
   describe "when namespace option is specified" do
@@ -189,14 +201,14 @@ describe JsRoutes, "options" do
       context "provided" do
         let(:_options) { { :default_url_options => { :optional_id => "12", :format => "json" } } }
         it "should use this opions to fill optional parameters" do
-          expect(evaljs("Routes.things_path()")).to eq(routes.things_path(:optional_id => 12, :format => "json"))
+          expect(evaljs("Routes.things_path()")).to eq(test_routes.things_path(:optional_id => 12, :format => "json"))
         end
       end
 
       context "not provided" do
         let(:_options) { { :default_url_options => { :format => "json" } } }
         it "breaks" do
-          expect(evaljs("Routes.foo_all_path()")).to eq(routes.foo_all_path(:format => "json"))
+          expect(evaljs("Routes.foo_all_path()")).to eq(test_routes.foo_all_path(:format => "json"))
         end
       end
     end
@@ -204,7 +216,15 @@ describe JsRoutes, "options" do
     context "with required route parts" do
       let(:_options) { {:default_url_options => {:inbox_id => "12"}} }
       it "should use this opions to fill optional parameters" do
-        expect(evaljs("Routes.inbox_messages_path()")).to eq(routes.inbox_messages_path(:inbox_id => 12))
+        expect(evaljs("Routes.inbox_messages_path()")).to eq(test_routes.inbox_messages_path(:inbox_id => 12))
+      end
+    end
+
+    context "when overwritten on JS level" do
+        let(:_options) { { :default_url_options => { :format => "json" } } }
+      it "uses JS defined value" do
+        evaljs("Routes.configure({default_url_options: {format: 'xml'}})")
+        expect(evaljs("Routes.inboxes_path()")).to eq(test_routes.inboxes_path(format: 'xml'))
       end
     end
   end
@@ -213,33 +233,33 @@ describe JsRoutes, "options" do
     context "with default option" do
       let(:_options) { Hash.new }
       it "should working in params" do
-        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true})")).to eq(routes.inbox_path(1, :trailing_slash => true))
+        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true})")).to eq(test_routes.inbox_path(1, :trailing_slash => true))
       end
 
       it "should working with additional params" do
-        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true, test: 'params'})")).to eq(routes.inbox_path(1, :trailing_slash => true, :test => 'params'))
+        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true, test: 'params'})")).to eq(test_routes.inbox_path(1, :trailing_slash => true, :test => 'params'))
       end
     end
 
     context "with default_url_options option" do
       let(:_options) { {:default_url_options => {:trailing_slash => true}} }
       it "should working" do
-        expect(evaljs("Routes.inbox_path(1, {test: 'params'})")).to eq(routes.inbox_path(1, :trailing_slash => true, :test => 'params'))
+        expect(evaljs("Routes.inbox_path(1, {test: 'params'})")).to eq(test_routes.inbox_path(1, :trailing_slash => true, :test => 'params'))
       end
 
       it "should remove it by params" do
-        expect(evaljs("Routes.inbox_path(1, {trailing_slash: false})")).to eq(routes.inbox_path(1))
+        expect(evaljs("Routes.inbox_path(1, {trailing_slash: false})")).to eq(test_routes.inbox_path(1))
       end
     end
 
     context "with disabled default_url_options option" do
       let(:_options) { {:default_url_options => {:trailing_slash => false}} }
       it "should not use trailing_slash" do
-        expect(evaljs("Routes.inbox_path(1, {test: 'params'})")).to eq(routes.inbox_path(1, :test => 'params'))
+        expect(evaljs("Routes.inbox_path(1, {test: 'params'})")).to eq(test_routes.inbox_path(1, :test => 'params'))
       end
 
       it "should use it by params" do
-        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true})")).to eq(routes.inbox_path(1, :trailing_slash => true))
+        expect(evaljs("Routes.inbox_path(1, {trailing_slash: true})")).to eq(test_routes.inbox_path(1, :trailing_slash => true))
       end
     end
   end
@@ -248,7 +268,7 @@ describe JsRoutes, "options" do
     context "with default option" do
       let(:_options) { Hash.new }
       it "should use snake case routes" do
-        expect(evaljs("Routes.inbox_path(1)")).to eq(routes.inbox_path(1))
+        expect(evaljs("Routes.inbox_path(1)")).to eq(test_routes.inbox_path(1))
         expect(evaljs("Routes.inboxPath")).to be_nil
       end
     end
@@ -258,8 +278,8 @@ describe JsRoutes, "options" do
       it "should generate camel case routes" do
         expect(evaljs("Routes.inbox_path")).to be_nil
         expect(evaljs("Routes.inboxPath")).not_to be_nil
-        expect(evaljs("Routes.inboxPath(1)")).to eq(routes.inbox_path(1))
-        expect(evaljs("Routes.inboxMessagesPath(10)")).to eq(routes.inbox_messages_path(:inbox_id => 10))
+        expect(evaljs("Routes.inboxPath(1)")).to eq(test_routes.inbox_path(1))
+        expect(evaljs("Routes.inboxMessagesPath(10)")).to eq(test_routes.inbox_messages_path(:inbox_id => 10))
       end
     end
   end
@@ -268,7 +288,7 @@ describe JsRoutes, "options" do
     context "with default option" do
       let(:_options) { Hash.new }
       it "should generate only path links" do
-        expect(evaljs("Routes.inbox_path(1)")).to eq(routes.inbox_path(1))
+        expect(evaljs("Routes.inbox_path(1)")).to eq(test_routes.inbox_path(1))
         expect(evaljs("Routes.inbox_url")).to be_nil
       end
     end
@@ -278,19 +298,19 @@ describe JsRoutes, "options" do
         let(:_options) { { :url_links => true, :default_url_options => {:host => "example.com"} } }
 
         it "uses the specified host, defaults protocol to http, defaults port to 80 (leaving it blank)" do
-          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com#{routes.inbox_path(1)}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com#{test_routes.inbox_path(1)}")
         end
 
         it "does not override protocol when specified in route" do
-          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{routes.new_session_path}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{test_routes.new_session_path}")
         end
 
         it "does not override host when specified in route" do
-          expect(evaljs("Routes.sso_url()")).to eq(routes.sso_url)
+          expect(evaljs("Routes.sso_url()")).to eq(test_routes.sso_url)
         end
 
         it "does not override port when specified in route" do
-          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{routes.portals_path}")
+          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{test_routes.portals_path}")
         end
       end
 
@@ -298,19 +318,19 @@ describe JsRoutes, "options" do
         let(:_options) { { :url_links => true, :default_url_options => {:host => "example.com", :protocol => "ftp"} } }
 
         it "uses the specified protocol and host, defaults port to 80 (leaving it blank)" do
-          expect(evaljs("Routes.inbox_url(1)")).to eq("ftp://example.com#{routes.inbox_path(1)}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("ftp://example.com#{test_routes.inbox_path(1)}")
         end
 
         it "does not override protocol when specified in route" do
-          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{routes.new_session_path}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{test_routes.new_session_path}")
         end
 
         it "does not override host when host is specified in route" do
-          expect(evaljs("Routes.sso_url()")).to eq("ftp://sso.example.com#{routes.sso_path}")
+          expect(evaljs("Routes.sso_url()")).to eq("ftp://sso.example.com#{test_routes.sso_path}")
         end
 
         it "does not override port when specified in route" do
-          expect(evaljs("Routes.portals_url()")).to eq("ftp://example.com:8080#{routes.portals_path}")
+          expect(evaljs("Routes.portals_url()")).to eq("ftp://example.com:8080#{test_routes.portals_path}")
         end
       end
 
@@ -318,49 +338,49 @@ describe JsRoutes, "options" do
         let(:_options) { { :url_links => true, :default_url_options => {:host => "example.com", :port => 3000} } }
 
         it "uses the specified host and port, defaults protocol to http" do
-          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com:3000#{routes.inbox_path(1)}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com:3000#{test_routes.inbox_path(1)}")
         end
 
         it "does not override protocol when specified in route" do
-          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com:3000#{routes.new_session_path}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com:3000#{test_routes.new_session_path}")
         end
 
         it "does not override host, protocol, or port when host is specified in route" do
-          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com:3000" + routes.sso_path)
+          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com:3000" + test_routes.sso_path)
         end
 
         it "does not override port when specified in route" do
-          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{routes.portals_path}")
+          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{test_routes.portals_path}")
         end
       end
 
       context "with camel_case option" do
         let(:_options) { { :camel_case => true, :url_links => true, :default_url_options => {:host => "example.com"} } }
         it "should generate path and url links" do
-          expect(evaljs("Routes.inboxUrl(1)")).to eq("http://example.com#{routes.inbox_path(1)}")
-          expect(evaljs("Routes.newSessionUrl()")).to eq("https://example.com#{routes.new_session_path}")
-          expect(evaljs("Routes.ssoUrl()")).to eq(routes.sso_url)
-          expect(evaljs("Routes.portalsUrl()")).to eq("http://example.com:8080#{routes.portals_path}")
+          expect(evaljs("Routes.inboxUrl(1)")).to eq("http://example.com#{test_routes.inbox_path(1)}")
+          expect(evaljs("Routes.newSessionUrl()")).to eq("https://example.com#{test_routes.new_session_path}")
+          expect(evaljs("Routes.ssoUrl()")).to eq(test_routes.sso_url)
+          expect(evaljs("Routes.portalsUrl()")).to eq("http://example.com:8080#{test_routes.portals_path}")
         end
       end
 
       context "with prefix option" do
         let(:_options) { { :prefix => "/api", :url_links => true, :default_url_options => {:host => 'example.com'} } }
         it "should generate path and url links" do
-          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com/api#{routes.inbox_path(1)}")
-          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com/api#{routes.new_session_path}")
-          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com/api#{routes.sso_path}")
-          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080/api#{routes.portals_path}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com/api#{test_routes.inbox_path(1)}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com/api#{test_routes.new_session_path}")
+          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com/api#{test_routes.sso_path}")
+          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080/api#{test_routes.portals_path}")
         end
       end
 
       context "with compact option" do
         let(:_options) { { :compact => true, :url_links => true, :default_url_options => {:host => 'example.com'} } }
         it "does not affect url helpers" do
-          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com#{routes.inbox_path(1)}")
-          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{routes.new_session_path}")
-          expect(evaljs("Routes.sso_url()")).to eq(routes.sso_url)
-          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{routes.portals_path}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("http://example.com#{test_routes.inbox_path(1)}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://example.com#{test_routes.new_session_path}")
+          expect(evaljs("Routes.sso_url()")).to eq(test_routes.sso_url)
+          expect(evaljs("Routes.portals_url()")).to eq("http://example.com:8080#{test_routes.portals_path}")
         end
       end
     end
@@ -385,23 +405,23 @@ describe JsRoutes, "options" do
         it "uses the current host" do
           expect(evaljs("Routes.inbox_path")).not_to be_nil
           expect(evaljs("Routes.inbox_url")).not_to be_nil
-          expect(evaljs("Routes.inbox_url(1)")).to eq("http://current.example.com#{routes.inbox_path(1)}")
-          expect(evaljs("Routes.inbox_url(1, { test_key: \"test_val\" })")).to eq("http://current.example.com#{routes.inbox_path(1, :test_key => "test_val")}")
-          expect(evaljs("Routes.new_session_url()")).to eq("https://current.example.com#{routes.new_session_path}")
-          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com#{routes.sso_path}")
+          expect(evaljs("Routes.inbox_url(1)")).to eq("http://current.example.com#{test_routes.inbox_path(1)}")
+          expect(evaljs("Routes.inbox_url(1, { test_key: \"test_val\" })")).to eq("http://current.example.com#{test_routes.inbox_path(1, :test_key => "test_val")}")
+          expect(evaljs("Routes.new_session_url()")).to eq("https://current.example.com#{test_routes.new_session_path}")
+          expect(evaljs("Routes.sso_url()")).to eq("http://sso.example.com#{test_routes.sso_path}")
 
         end
 
         it "uses host option as an argument" do
-          expect(evaljs("Routes.portals_url({host: 'another.com'})")).to eq(routes.portals_url(host: 'another.com'))
+          expect(evaljs("Routes.portals_url({host: 'another.com'})")).to eq(test_routes.portals_url(host: 'another.com'))
         end
 
         it "uses port option as an argument" do
-          expect(evaljs("Routes.portals_url({host: 'localhost', port: 8080})")).to eq(routes.portals_url(host: 'localhost', port: 8080))
+          expect(evaljs("Routes.portals_url({host: 'localhost', port: 8080})")).to eq(test_routes.portals_url(host: 'localhost', port: 8080))
         end
 
         it "uses protocol option as an argument" do
-          expect(evaljs("Routes.portals_url({host: 'localhost', protocol: 'https'})")).to eq(routes.portals_url(protocol: 'https', host: 'localhost'))
+          expect(evaljs("Routes.portals_url({host: 'localhost', protocol: 'https'})")).to eq(test_routes.portals_url(protocol: 'https', host: 'localhost'))
         end
       end
     end
@@ -411,8 +431,8 @@ describe JsRoutes, "options" do
     let(:_options) { { :compact => true } }
     it "removes _path suffix from path helpers" do
       expect(evaljs("Routes.inbox_path")).to be_nil
-      expect(evaljs("Routes.inboxes()")).to eq(routes.inboxes_path())
-      expect(evaljs("Routes.inbox(2)")).to eq(routes.inbox_path(2))
+      expect(evaljs("Routes.inboxes()")).to eq(test_routes.inboxes_path())
+      expect(evaljs("Routes.inbox(2)")).to eq(test_routes.inbox_path(2))
     end
 
     context "with url_links option" do
@@ -424,8 +444,8 @@ describe JsRoutes, "options" do
 
       let(:_options) { { :compact => true, :url_links => true, default_url_options: {host: 'localhost'} } }
       it "should not strip urls" do
-        expect(evaljs("Routes.inbox(1)")).to eq(routes.inbox_path(1))
-        expect(evaljs("Routes.inbox_url(1)")).to eq("http://localhost#{routes.inbox_path(1)}")
+        expect(evaljs("Routes.inbox(1)")).to eq(test_routes.inbox_path(1))
+        expect(evaljs("Routes.inbox_url(1)")).to eq("http://localhost#{test_routes.inbox_path(1)}")
       end
     end
   end
@@ -436,7 +456,7 @@ describe JsRoutes, "options" do
       expect {
         expect(evaljs("Routes.inbox_message_path({inbox_id: 1, id: 2, _options: true})")).to eq("")
       }.to raise_error(js_error_class)
-      expect(evaljs("Routes.inbox_message_path({inbox_id: 1, id: 2, __options__: true})")).to eq(routes.inbox_message_path(inbox_id: 1, id: 2))
+      expect(evaljs("Routes.inbox_message_path({inbox_id: 1, id: 2, __options__: true})")).to eq(test_routes.inbox_message_path(inbox_id: 1, id: 2))
     end
   end